Upgrade i3-gaps-next to version 4.15.0.1.20180920 (52240a0)
i3-gaps-next commit 6a1f653 references the g_utf8_make_valid
function, which was introduced in GLib 2.52. The FreeBSD ports
tree currently has GLib 2.50, so we need to create a port of
our own for 2.52 (see commit 3183952) and depend on it instead
until the mainline port has been updated to 2.52.